Time: 0 ms (100.00%) | Memory: 21.5 MB (8.41%) - LeetSync

+++ 2614-maximum-count-of-positive-integer-and-negative-integer/maximum-count-of-positive-integer-and-negative-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int maximumCount(vector<int>& nums) {
-        int n=nums.size();
+        size_t n=nums.size();
         int odd=0;
         int even=0;
         for(auto x:nums){
